scripts/requirements-meeting-notes.txt
# Dependencies for Meeting Notes Auto-Import
openai>=1.0.0
supabase>=2.0.0
python-dotenv>=1.0.0
watchdog>=3.0.0
